Move the rest of logic to useEditor
