+++ src/application/services/useNote.ts
@@ -49,7 +49,7 @@ interface UseNoteComposableState {
   /**
    * Creates/updates the note
    */
-  save: (content: NoteContent, parentId: NoteId | null) => Promise<void>;
+  save: (content: NoteContent, parentId: NoteId | null | undefined) => Promise<void>;
 
   /**
    * Load note by custom hostname
@@ -139,7 +139,7 @@ export default function (options: UseNoteComposableOptions): UseNoteComposableSt
    * @param content - Note content (Editor.js data)
    * @param parentId - Id of the parent note. If null, then it's a root note
    */
-  async function save(content: NoteContent, parentId: NoteId | null): Promise<void> {
+  async function save(content: NoteContent, parentId: NoteId | undefined): Promise<void> {
     if (note.value === null) {
       throw new Error('Note is not loaded yet');
     }
