+++ src/infrastructure/note.repository.ts
@@ -71,11 +71,11 @@ export default class NoteRepository implements NoteRepositoryInterface {
    * Creates a new note
    *
    * @param content - Note content (Editor.js data)
-   * @param parentId - Id of the parent note. If null, then it's a root note
+   * @param parentId - Id of the parent note. If undefined, then it's a root note
    *
    * @todo API should return Note
    */
-  public async createNote(content: NoteContent, parentId: NoteId | null): Promise<Note> {
+  public async createNote(content: NoteContent, parentId?: NoteId): Promise<Note> {
     const response = await this.transport.post<{ id: NoteId }>('/note', {
       content,
       parentId,
