+++ src/presentation/components/editor/Editor.vue
@@ -3,15 +3,20 @@
 </template>
 
 <script setup lang="ts">
-import { ref } from 'vue';
-import { type OutputData, type API } from '@editorjs/editorjs';
+import { type OutputData } from '@editorjs/editorjs';
 import { useEditor } from '@/application/services/useEditor';
 
 /**
  * Define the props for the component
  */
 const props = defineProps<{
+  /**
+   * Editor content
+   */
   content?: OutputData;
+  /**
+   * True if editor content is not editable
+   */
   readOnly?: boolean;
 }>();
 
@@ -19,65 +24,11 @@ const emit = defineEmits<{
   change: [data: OutputData];
 }>();
 
-/**
- * Attribute containing is-empty state.
- * It is updated on every change of the editor
- */
-const isEmpty = ref(true);
-
-/**
- * Checks if the editor is empty
- * Uses EditorJS API:
- *  - blocks.getById()
- *  - block.isEmpty()
- *
- * @todo implement "isEmpty" method in the EditorJS API
- *
- * @param data - saved data
- * @param api - EditorJS API
- */
-function checkIsEmpty(data: OutputData, api: API): boolean {
-  const blockIds = data.blocks.map((block) => block.id);
-
-  return blockIds.reduce((acc, id) => {
-    if (id === undefined) {
-      return acc;
-    }
-
-    const block = api.blocks.getById(id);
-
-    if (block) {
-      return acc && block.isEmpty;
-    }
-
-    return acc;
-  }, true);
-}
-
-/**
- * Function called on every change of the editor
- *
- * @param api - EditorJS API
- */
-async function onChange(api: API): Promise<void> {
-  const data = await api.saver.save();
-
-  /**
-   * Update the isEmpty attribute
-   */
-  isEmpty.value = checkIsEmpty(data, api);
-
-  /**
-   * Change model value
-   */
-  emit('change', data);
-}
-
-useEditor({
+const { isEmpty } = useEditor({
   id: 'editorjs',
   content: props.content,
   isReadOnly: props.readOnly,
-  onChange,
+  onChange: (data) => emit('change', data),
 });
 
 defineExpose({
